Move `initStarterColors()` out of `BattleScene`
